+++ lib/output/amqp_0_9.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 Ashley Jeffs
+// Copyright (c) 2014 Ashley Jeffs
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -30,56 +30,37 @@ import (
 //------------------------------------------------------------------------------
 
 func init() {
-	Constructors[TypeRedisHash] = TypeSpec{
-		constructor: NewRedisHash,
+	Constructors[TypeAMQP09] = TypeSpec{
+		constructor: NewAMQP09,
 		description: `
-Sets Redis hash objects using the HMSET command.
+Sends messages to an AMQP (0.91) exchange. AMQP is a messaging protocol used by
+various message brokers, including RabbitMQ. The metadata from each message are
+delivered as headers.
 
-The field ` + "`key`" + ` supports
-[interpolation functions](../config_interpolation.md#functions) evaluated per
-message of a batch, allowing you to create a unique key for each message.
+It's possible for this output type to create the target exchange by setting
+` + "`exchange_declare.enabled` to `true`" + `, if the exchange already exists
+then the declaration passively verifies that the settings match.
 
-The field ` + "`fields`" + ` allows you to specify an explicit map of field
-names to interpolated values, also evaluated per message of a batch:
+Exchange type options are: direct|fanout|topic|x-custom
 
-` + "```yaml" + `
-redis_hash:
-  url: tcp://localhost:6379
-  key: ${!json_field:id}
-  fields:
-    topic: ${!metadata:kafka_topic}
-    partition: ${!metadata:kafka_partition}
-    content: ${!json_field:document.text}
-` + "```" + `
+TLS is automatic when connecting to an ` + "`amqps`" + ` URL, but custom
+settings can be enabled in the ` + "`tls`" + ` section.
 
-If the field ` + "`walk_metadata`" + ` is set to ` + "`true`" + ` then Benthos
-will walk all metadata fields of messages and add them to the list of hash
-fields to set.
-
-If the field ` + "`walk_json_object`" + ` is set to ` + "`true`" + ` then
-Benthos will walk each message as a JSON object, extracting keys and the string
-representation of their value and adds them to the list of hash fields to set.
-
-The order of hash field extraction is as follows:
-
-1. Metadata (if enabled)
-2. JSON object (if enabled)
-3. Explicit fields
-
-Where latter stages will overwrite matching field names of a former stage.`,
+The field 'key' can be dynamically set using function interpolations described
+[here](../config_interpolation.md#functions).`,
 	}
 }
 
 //------------------------------------------------------------------------------
 
-// NewRedisHash creates a new RedisHash output type.
-func NewRedisHash(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
-	rhash, err := writer.NewRedisHash(conf.RedisHash, log, stats)
+// NewAMQP09 creates a new AMQP output type.
+func NewAMQP09(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
+	a, err := writer.NewAMQP(conf.AMQP09, log, stats)
 	if err != nil {
 		return nil, err
 	}
 	return NewWriter(
-		TypeRedisHash, rhash, log, stats,
+		TypeAMQP09, a, log, stats,
 	)
 }