+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- New `amqp_0_9` output which replaces the now deprecated `amqp` output.
+
 ## 3.3.1 - 2019-10-21
 
 ### Fixed